fix: 結果ページが表示されないバグを修正
- createShuwaDetailHTML関数を独立したコンポーネントに分離
- result/show-result.tsがlearn/shuwa-item.tsから安全にimportできるよう改善
- shuwa-item.tsの実行時エラー（存在しない.shuwa-items要素への参照）を解決
